rovio node skeleton
